requirements.txt
pandas
openpyxl==3.1.2
pywin32==307
#python-dotenv==1.0.0
python-dotenv==1.0.1
matplotlib>=3.9.0
seaborn>=0.13.0

